+++ database/migrations/20241010182000_add_timezone_to_sessions_table.php
@@ -1,0 +1,24 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Capsule\Manager as DB;
+
+class AddTimezoneToSessionsTable extends Migration
+{
+    public function up()
+    {
+        DB::table('sessions')->truncate();
+
+        $this->schema->table('sessions', function (Blueprint $table) {
+            $table->string('timezone')->index();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('sessions', function (Blueprint $table) {
+            $table->dropColumn('timezone');
+        });
+    }
+}