+++ src/Movim/Daemon/Core.php
@@ -125,7 +125,6 @@ class Core implements MessageComponentInterface
 
             if (!array_key_exists($sid, $this->sessions)) {
                 $language = $this->getLanguage($conn);
-                $offset = $this->getOffset($conn);
 
                 $this->sessions[$sid] = new Session(
                     $this->loop,
@@ -134,7 +133,6 @@ class Core implements MessageComponentInterface
                     config('daemon.port'),
                     $this->key,
                     $language,
-                    $offset,
                     config('daemon.verbose'),
                     config('daemon.debug')
                 );
@@ -274,12 +272,6 @@ class Core implements MessageComponentInterface
     {
         $languages = $conn->httpRequest->getHeader('Accept-Language');
         return (is_array($languages) && !empty($languages)) ? $languages[0] : false;
-    }
-
-    private function getOffset(ConnectionInterface $conn)
-    {
-        parse_str($conn->httpRequest->getUri()->getQuery(), $arr);
-        return (isset($arr['offset'])) ? invertSign(((int)$arr['offset']) * 60) : 0;
     }
 
     private function getPath(ConnectionInterface $conn)